refactor: renamed rxtree classifier operator
